+++ src/renderer/components/dialogs/ViewGroup.tsx
@@ -21,7 +21,11 @@ import { ViewProfileInner } from './ViewProfile'
 import { ScreenContext, useTranslationFunction } from '../../contexts'
 import { useState, useContext, useEffect, useCallback, useMemo } from 'react'
 import React from 'react'
-import { Avatar, avatarInitial } from '../Avatar'
+import {
+  Avatar,
+  avatarInitial,
+  ClickForFullscreenAvatarWrapper,
+} from '../Avatar'
 import { runtime } from '../../runtime'
 import { DeltaInput } from '../Login-Styles'
 import { ipcBackend } from '../../ipc'
@@ -202,13 +206,15 @@ function ViewGroupInner(props: {
           <div className={Classes.DIALOG_BODY}>
             <Card>
               <div className='group-settings-container'>
-                <Avatar
-                  displayName={groupName}
-                  avatarPath={groupImage}
-                  isVerified={chat.isProtected}
-                  color={chat.color}
-                  large
-                />
+                <ClickForFullscreenAvatarWrapper filename={groupImage}>
+                  <Avatar
+                    displayName={groupName}
+                    avatarPath={groupImage}
+                    isVerified={chat.isProtected}
+                    color={chat.color}
+                    large
+                  />
+                </ClickForFullscreenAvatarWrapper>
                 <p className='group-name' style={{ marginLeft: '17px' }}>
                   {groupName}
                 </p>