+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - Add logging in by pressing enter in AccountSetupScreen
 - Fix blob width for html messages
 - Fix usage of wrong translation key chat_no_contact_requests
+- Fullscreen view for group avatar and own avatar in settings
 
 ## [1.26.0] - 2021-12-15